// src/lib/supabase/types.ts
export interface Database {
// ...
  public: {
    Tables: {
      players: {
        Row: {
          birthdate: string
          cellphone: string | null
          created_at: string | null
          dni: number
          email: string | null
          id: string
          lastname: string
          name: string
          observations: string | null
        }
        Insert: {
          birthdate: string
          cellphone?: string | null
          created_at?: string | null
          dni: number
          email?: string | null
          id?: string
          lastname?: string
          name?: string
          observations?: string | null
        }
        Update: {
          birthdate?: string
          cellphone?: string | null
          created_at?: string | null
          dni?: number
          email?: string | null
          id?: string
          lastname?: string
          name?: string
          observations?: string | null
        }
        Relationships: []
      }
      players_sports: {
        Row: {
          created_at: string | null
          federated: boolean
          player_id: string
          sport_id: string
        }
        Insert: {
          created_at?: string | null
          federated?: boolean
          player_id: string
          sport_id: string
        }
        Update: {
          created_at?: string | null
          federated?: boolean
          player_id?: string
          sport_id?: string
        }
        Relationships: [
          {
            foreignKeyName: "players_sports_player_id_fkey"
            columns: ["player_id"]
            referencedRelation: "players"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "players_sports_sport_id_fkey"
            columns: ["sport_id"]
            referencedRelation: "sports"
            referencedColumns: ["id"]
          }
        ]
      }
// ...
      sports: {
        Row: {
          created_at: string | null
          id: string
          name: string
        }
        Insert: {
          created_at?: string | null
          id?: string
          name: string
        }
        Update: {
          created_at?: string | null
          id?: string
          name?: string
        }
        Relationships: []
      }
// ...
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
// ...
}